Purchase - add links to TankerLoad, Haul & Delivery | add Haul figures & details | change integer to decimal and add formats

+++ database/migrations/2020_12_26_002314_create_purchase_details_table.php
@@ -15,8 +15,8 @@ class CreatePurchaseDetailsTable extends Migration
     {
         Schema::create('purchase_details', function (Blueprint $table) {
             $table->id();
-            $table->integer('quantity')->nullable();
-            $table->integer('unit_price')->nullable();
+            $table->decimal('quantity', 8, 0)->nullable();
+            $table->decimal('unit_price', 10, 3)->nullable();
             $table->longText('remarks')->nullable();
             $table->bigInteger('purchase_id');
             $table->bigInteger('product_id')->nullable();
